Remove go-bindata (#1675)
This (fork of) go-bindata is not awesome.

The project is not maintained anymore and has a great bunch of issues.
Currently there are 50 unresolved issue tickets and 18 unmerge pull
requests. Many of the issues make this tool very annoying to use, but
would be easy to fix if it was maintained.

An example issue that is very easy to fix is
https://github.com/jteeuwen/go-bindata/issues/148

See also https://github.com/jteeuwen/go-bindata/issues/165
Apparently the maintainer has disappeared from earth's surface two
years ago. (Let's hope that's not actually true.)